Refactor product imports to use interface files
Updated imports in favorites and shopping cart components to reference '../product/interface' instead of '../product/product'. Renamed 'favorite.ts' to 'interface.ts' in the favorites component for consistency.
