+++ components/favorites/interface.ts
@@ -1,4 +1,4 @@
-import { Product } from "../product/product";
+import { Product } from "../product/interface";
 
 export interface Favorite {
   id: string;
