+++ components/shopping-cart/shopping-cart.ts
@@ -1,4 +1,4 @@
-import { Product } from "../product/product";
+import { Product } from "../product/interface";
 
 export interface ShoppingCart {
   id: string;
